+++ projects/client/src/pages/user/Cart.jsx
@@ -7,11 +7,10 @@ import { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { countItem } from "../../redux/cartSlice";
 import pin from "../../assets/images/pin.png"
-import { useNavigate } from "react-router-dom";
 import { checkDiscount, countDiscount, rupiah, getTotalPrice, getTotalWeight } from "../../function";
+import CheckoutInformation from "../../component/CheckoutInformation";
 
 export default function Cart() {
-  const Navigate = useNavigate()
   const user = useSelector((state) => state.userSlice);
   const branchId = useSelector((state) => state.branchSlice.branchId);
   const token = localStorage.getItem("token")
@@ -26,7 +25,6 @@ export default function Cart() {
               'Authorization': `Bearer ${token}`
           }
       });
-
       dispatch(
         countItem({
           count: response.data.data,
@@ -46,7 +44,6 @@ export default function Cart() {
             }
         });
         const cartData = response.data.results;
-
         setCarts(cartData);
       } catch (error) {
         toast.error(error.response.data.message);
@@ -112,9 +109,7 @@ export default function Cart() {
         
           <form className="mt-12 lg:grid lg:grid-cols-12 lg:items-start lg:gap-x-12 xl:gap-x-16">
             <section aria-labelledby="cart-heading" className="lg:col-span-7">
-              <h2 id="cart-heading" className="sr-only">
-                Items in your shopping cart
-              </h2>
+              <h2 id="cart-heading" className="sr-only">Items in your shopping cart</h2>
 
               <ul
                 role="list"
@@ -135,10 +130,7 @@ export default function Cart() {
                         <div>
                           <div className="flex justify-between">
                             <h3 className="text-sm">
-                              <a
-                                href={`/product/${item.id_inventory}`}
-                                className="font-medium text-gray-700 hover:text-gray-800"
-                              >
+                              <a href={`/product/${item.id_inventory}`} className="font-medium text-gray-700 hover:text-gray-800">
                                 {item.product_name}
                               </a>
                             </h3>
@@ -158,54 +150,35 @@ export default function Cart() {
                           : 
                           <p className="mt-1 text-sm font-medium text-gray-900">
                              {rupiah(item.product_price)}
-                          </p>
-                          }
+                          </p>}
                           {checkDiscount(item) == 'bonus_qty' &&
                             <p className="mt-1 text-sm font-medium text-gray-900">
                               Bonus item: {item.bonus_qty} pcs
-                            </p>
-                          }
-                          
+                            </p>}
                         </div>
-
                         <div className="mt-4 sm:mt-0 sm:pr-9">
                           <div class="flex items-center">
                             <button type="button" class="flex-none mr-3">
                               <MinusIcon onClick={() => updateQty(item.id, -1)} className="h-3 w-3 fill-gray-500 hover:fill-gray-700 outline outline-offset-2 outline-1" aria-hidden="true"/>
                             </button>
-                            <div class="flex-none">
-                              {item.product_qty}
-                            </div>
+                            <div class="flex-none"> {item.product_qty}</div>
                             <button type="button" class="flex-none ml-3">
                               <PlusIcon onClick={() => updateQty(item.id, 1)}className="h-3 w-3 fill-gray-500 hover:fill-gray-700 outline outline-offset-2 outline-1" aria-hidden="true"/>
                             </button>
                           </div>
 
                           <div className="absolute top-0 right-0">
-                            <button
-                              type="button"
-                              className="-m-2 inline-flex p-2 text-gray-400 hover:text-gray-500"
-                              onClick={() => deleteItem(item.id)}
-                            >
+                            <button type="button" className="-m-2 inline-flex p-2 text-gray-400 hover:text-gray-500" onClick={() => deleteItem(item.id)}>
                               <span className="sr-only">Remove</span>
-                              <XMarkIcon
-                                className="h-5 w-5"
-                                aria-hidden="true"
-                              />
+                              <XMarkIcon className="h-5 w-5" aria-hidden="true"/>
                             </button>
                           </div>
                         </div>
                       </div>
 
                       <p className="mt-4 flex space-x-2 text-sm text-gray-700">
-                        <CheckIcon
-                        className="h-5 w-5 flex-shrink-0 text-green-500"
-                        aria-hidden="true"
-                        />
-
-                        <span>
-                            Stock: {item.stock}
-                        </span>
+                        <CheckIcon className="h-5 w-5 flex-shrink-0 text-green-500" aria-hidden="true"/>
+                        <span>Stock: {item.stock}</span>
                       </p>
                     </div>
                   </li>
@@ -213,45 +186,7 @@ export default function Cart() {
               </ul>
             </section>
 
-            {/* Order summary */}
-            <section
-              aria-labelledby="summary-heading"
-              className="mt-16 rounded-lg bg-gray-50 px-4 py-6 sm:p-6 lg:col-span-5 lg:mt-0 lg:p-8"
-            >
-              <h2
-                id="summary-heading"
-                className="text-lg font-medium text-gray-900"
-              >
-                Order Total
-              </h2>
-
-              <dl className="mt-6 space-y-4">
-                <div className="flex items-center justify-between">
-                    <dt className="text-sm text-gray-600">Total Weight</dt>
-                    <dd className="text-sm font-medium text-gray-900">{getTotalWeight(carts)} gr</dd>
-                </div>
-                <div className="flex items-center justify-between border-t border-gray-200 pt-4">
-                  <dt className="text-base font-medium text-gray-900">
-                    Total Price
-                  </dt>
-                  <dd className="text-base font-medium text-gray-900">
-                    {rupiah(getTotalPrice(carts))}
-                  </dd>
-                </div>
-              </dl>
-
-              {carts.length > 0 &&
-                <div className="mt-6">
-                  <button
-                    onClick={()=> Navigate('/order')}
-                    type="submit"
-                    className="w-full rounded-md border border-transparent bg-green-600 py-3 px-4 text-base font-medium text-white shadow-sm hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:ring-offset-gray-50"
-                  >
-                    Checkout
-                  </button>
-                </div>
-              }
-            </section>
+            <CheckoutInformation carts={carts} />
           </form>
         </div>
         <Toaster/>
