+++ projects/server/src/controllers/profileController.js
@@ -25,9 +25,10 @@ module.exports = {
     },
 
     editProfile: async (req, res) => {
-        try {          
+        try {  
+            var isDate = function(date) { return (new Date(date) !== "Invalid Date") && !isNaN(new Date(date));}        
             let {name, gender, email, birthdate, prevEmail} = req.body;
-            if(!birthdate) birthdate = null
+            if(isDate(birthdate)==false) birthdate = null
             if (!email) return res.status(400).send({isError: true, message: "Please fill the required field (email)"});
             if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)) return res.status(400).send({isError: true, message: "Invalid email format"});
             let findEmail = await users.findOne({ where: { email: email } });
@@ -37,13 +38,21 @@ module.exports = {
 
             if(req.file != undefined){
                 let imageUrl = process.env.API_URL + "/media/profiles/" + req.file.filename;
-                if(!birthdate) await users.update({name, gender, email, birthdate: null, profile_picture: imageUrl}, {where: {id: req.params.id}});
-                else await users.update({name, gender, email, birthdate, profile_picture: imageUrl}, {where: {id: req.params.id}});
+                if(!birthdate){
+                    console.log("satu")
+                    await users.update({name, gender, email, birthdate: null, profile_picture: imageUrl}, {where: {id: req.params.id}});
+                    console.log("dua")
+                }
+                else{
+                    console.log("tiga")
+                    await users.update({name, gender, email, birthdate, profile_picture: imageUrl}, {where: {id: req.params.id}});}
             }else{
-                if(birthdate) await users.update({name, gender, email, birthdate: null}, {where: {id: req.params.id}});
-                else await users.update({name, gender, email, birthdate}, {where: {id: req.params.id}});
+                console.log("empat")
+                if(!birthdate) {await users.update({name, gender, email, birthdate: null}, {where: {id: req.params.id}});}
+                else{ await users.update({name, gender, email, birthdate}, {where: {id: req.params.id}}); }
             }
-
+            
+            
             res.status(200).send({isError: false, message: "Profile updated", data: req.body});
           
         } catch (error) {
